requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
aiomqtt==2.3.0
aiocoap==0.4.12
zeroconf==0.136.0
pydantic==2.9.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
pyyaml==6.0.2
pycryptodome==3.21.0
cryptography==43.0.3
python-crontab==3.2.0
websockets==13.1
httpx==0.27.2
